Cast osmajorrelease to string
Depending on the version of Salt we are running or the release it is
running on, the `osmajorrelease` grain may return a string or an
integer. To be on the safe side we cast the value to a string so that it
will always correctly match the dict key values.