Update centos7 templates to be consistent with upstream
Most of these changes are just to commented sections of the config files
so have no real impact on the configuration of ssh. However the extra
noise these changes add makes it harder to see where "real" changes are
being made. If we can keep these config files consistent with upstream,
then the changes to actual settings will be more visible.